Optimized inHistory method
inHistory now checks if it has seen the configuration of boxes and if it does find it checks if the player can move itself to the position where it was previously observed.

If it finds that the configuration is new and/or the player cant move to the
a previous position it will add the players current position under that particular configuration of boxes.